requirements.txt
Jupyter
Pillow
pytesseract
opencv-python
matplotlib
imageio